refactor: вынести express server в отдельный файл expressServer.js
