+++ README.md
@@ -9,39 +9,41 @@
 5. очистить cloudinary каждые 2 часа
 6. удалить пользователя
 7.  добавить пакеты безопасности. В частности, host-csrf? урок Using EJS In A Database Application, week15
-8.  Рефактор app.js - все роутеры в отдельный файл (Богдан 154, 155)
-9.  переписать почту с тестовой на желаемую GMAIL
-10. добавить тесты puppeteer
-11. создать запись админа, чтобы чистить базу данных?
-12. двухфакторная аутентификация
-13. выходить из учетки при изменении пароля или при изменениия email
-14. app.use(fileUpload) переписать этот middleware только для конвертации и вынести в папку middleware
+8.  переписать почту с тестовой на желаемую GMAIL
+9.  добавить тесты puppeteer
+10. создать запись админа, чтобы чистить базу данных?
+11. двухфакторная аутентификация
+12. выходить из учетки при изменении пароля или при изменениия email
+13. app.use(fileUpload) переписать этот middleware только для конвертации и вынести в папку middleware
+14. добавить эндпойнт для проверки Render.com, запущен ли сервер (перезапуск)
+15. passport?
+16. настроить часовые пояса на tasks.js и accessLog.js
     ФРОНТ:
-15. редирект на страницу с результатами (кнопка или ссылка "Скачать" и "Конвертировать другой архив")
-16. перетащить архив в input
-17. добавить функцию выбрать язык?
-18. убрать высоту лого в header
-19. Галочка Запомнить меня при авторизации
-20. На странице регистрации и авторизации поднять повыше форму (на уровень глаз)
-21. Переписать на React
-22. валидация пароля и email? Пароль должен удовлетворять требованиям (минимум 6 символов, заглавные и прописные, символы, без пробелов)
-23. вынести в отдельную функцию fetchData
-24. одна функция запроса на бэк getData
-25. переписать register.js и login.js в одну функцию?
-26. выровнять строки в таблице задач
-27. если на любой странице html ошибка, выдать окно или блок с сообщением
-28. сохранять файл под собственным именем
-29. удалить customError в связи с неиспользованием?
-30. убрать пустой треугольник в alert-msg на странице login.html (регистрация тоже?)
-31. доработать ссылку для подтверждения email ${origin}/user/verify-email (origin и эндпойнт)
-32. добавить аватарку - первая буква имени
-33. рефактор на одну функцию user.js и register.js
-34. рефактор user.js и history.js - повторяются операции (в том числе window.assign)
-35. добавить в регистрацию "повторите пароль"
-36. при изменении пароля или почты отправить на старую почту письмо "Сделан запрос на изменения почты/пароля. Если это были не Вы, измените пароль"
-37. добавить обработку ссылок с истекшим сроком (verify-email, update-pasword, register)
-38. добавить для пользователя Изменить пароль в ЛК
-39. блокировать таблицу history.js на время загрузки задач и на время удаления задачи
+17. редирект на страницу с результатами (кнопка или ссылка "Скачать" и "Конвертировать другой архив")
+18. перетащить архив в input
+19. добавить функцию выбрать язык?
+20. убрать высоту лого в header
+21. Галочка Запомнить меня при авторизации
+22. На странице регистрации и авторизации поднять повыше форму (на уровень глаз)
+23. Переписать на React
+24. валидация пароля и email? Пароль должен удовлетворять требованиям (минимум 6 символов, заглавные и прописные, символы, без пробелов)
+25. вынести в отдельную функцию fetchData
+26. одна функция запроса на бэк getData
+27. переписать register.js и login.js в одну функцию?
+28. выровнять строки в таблице задач
+29. если на любой странице html ошибка, выдать окно или блок с сообщением
+30. сохранять файл под собственным именем
+31. удалить customError в связи с неиспользованием?
+32. убрать пустой треугольник в alert-msg на странице login.html (регистрация тоже?)
+33. доработать ссылку для подтверждения email ${origin}/user/verify-email (origin и эндпойнт)
+34. добавить аватарку - первая буква имени
+35. рефактор на одну функцию user.js и register.js
+36. рефактор user.js и history.js - повторяются операции (в том числе window.assign)
+37. добавить в регистрацию "повторите пароль"
+38. при изменении пароля или почты отправить на старую почту письмо "Сделан запрос на изменения почты/пароля. Если это были не Вы, измените пароль"
+39. добавить обработку ссылок с истекшим сроком (verify-email, update-pasword, register)
+40. добавить для пользователя Изменить пароль в ЛК
+41. блокировать таблицу history.js на время загрузки задач и на время удаления задачи
 
 Использую
